minor fixes - loggin in GlobalExceptionHandler, rename typos in files

+++ spring-backend/src/main/kotlin/com/pavelryzh/firewallus/rule/service/RuleService.kt
@@ -49,6 +49,7 @@ class RuleService(private val ruleRepo: RuleRepository, private val eventPublish
     }
     @Transactional
     fun deleteRule(id: Int) {
+        if (!ruleRepo.existsById(id)) throw RuleNotFoundException(id)
         ruleRepo.deleteById(id)
         publishDeletedEvent(id)
     }
